feat: display more metadata
Metadata are displayed in a table with scroll.

The following metadata are the first items when available:

 - platformSerialIdentifier
 - instrument
 - productType
 - processingLevel
 - sensorMode
 - startTimeFromAscendingNode
 - completionTimeFromAscendingNode
 - orbitDirection
 - orbitNumber
 - cloudCover

Other metadata are listed in alphabetic order below this list.

Metadata labels are constructed from metadata’s name splited on upper
letters (split camel case) in order to be generic.

Metadata whose values are different from string, number or list of
strings are not displayed.

ref: #10